+++ crates/catalog/src/catalogs/mod.rs
@@ -1,2 +1,1 @@
 pub mod embucket;
-pub mod s3tables;
