--- crates/catalog/src/catalogs/s3tables/mod.rs
@@ -1,1 +1,0 @@
-pub mod iceberg_catalog;
