+++ crates/executor/Cargo.toml
@@ -28,9 +28,6 @@ sqlparser = { git = "https://github.com/Embucket/datafusion-sqlparser-rs.git", r
   "visitor",
 ] }
 iceberg-rust = { workspace = true }
-iceberg-rust-spec = { workspace = true }
-iceberg-s3tables-catalog = { workspace = true }
-aws-sdk-s3tables = "1.45.0"
 error-stack-trace = { path = "../error-stack-trace" }
 error-stack = { path = "../error-stack" }
